// src/app/core/auto_api/backend/model/enrollmentPeriodRequest.ts
/**
 * Api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { EnrollmentPeriodAcademicDegree } from './enrollmentPeriodAcademicDegree';

export interface EnrollmentPeriodRequest { 
    academicYearPrefix?: string;
    academicYearStart?: number;
    academicYearEnd?: number;
    academicDegree?: EnrollmentPeriodAcademicDegree;
    startTime?: Date;
    isSendingResumesFreely?: number;
}